requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.txt
#
asgiref==3.4.1
    # via
    #   -r requirements.txt
    #   django
boto3==1.18.58
    # via -r requirements.txt
botocore==1.21.58
    # via
    #   -r requirements.txt
    #   boto3
    #   s3transfer
click==8.0.3
    # via
    #   -r requirements.txt
    #   pip-tools
django==3.2.8
    # via
    #   -r requirements.txt
    #   django-cors-headers
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.10.0
    # via -r requirements.txt
django-storages==1.12
    # via -r requirements.txt
djangorestframework==3.12.4
    # via
    #   -r requirements.txt
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.8.0
    # via -r requirements.txt
gunicorn==20.1.0
    # via -r requirements.txt
jmespath==0.10.0
    # via
    #   -r requirements.txt
    #   boto3
    #   botocore
pep517==0.11.0
    # via
    #   -r requirements.txt
    #   pip-tools
pillow==8.3.2
    # via -r requirements.txt
pip-tools==6.3.1
    # via -r requirements.txt
psycopg2==2.9.1
    # via -r requirements.txt
pyjwt==2.1.0
    # via
    #   -r requirements.txt
    #   djangorestframework-simplejwt
python-dateutil==2.8.2
    # via
    #   -r requirements.txt
    #   botocore
pytz==2021.3
    # via
    #   -r requirements.txt
    #   django
s3transfer==0.5.0
    # via
    #   -r requirements.txt
    #   boto3
six==1.16.0
    # via
    #   -r requirements.txt
    #   python-dateutil
sqlparse==0.4.2
    # via
    #   -r requirements.txt
    #   django
tomli==1.2.1
    # via
    #   -r requirements.txt
    #   pep517
urllib3==1.26.7
    # via
    #   -r requirements.txt
    #   botocore
wheel==0.37.0
    # via pip-tools
yapf==0.31.0
    # via -r requirements.txt

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
